# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeLab – LeetCode-style competitive coding platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
  solvedCount: number;
  rank: number;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface TestCase {
  input: string;
  expectedOutput: string;
  isHidden?: boolean;
}

export type ProblemDifficulty = typeof ProblemDifficulty[keyof typeof ProblemDifficulty];


export const ProblemDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Problem {
  id: number;
  title: string;
  slug: string;
  difficulty: ProblemDifficulty;
  description: string;
  /** @nullable */
  constraints?: string | null;
  /** @nullable */
  hints?: string | null;
  tags: string[];
  /** @nullable */
  starterCode?: string | null;
  testCases?: TestCase[];
  createdAt: string;
  solvedCount: number;
  acceptanceRate: number;
  /** @nullable */
  userStatus?: string | null;
}

export type ProblemInputDifficulty = typeof ProblemInputDifficulty[keyof typeof ProblemInputDifficulty];


export const ProblemInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface ProblemInput {
  title: string;
  difficulty: ProblemInputDifficulty;
  description: string;
  constraints?: string;
  hints?: string;
  tags: string[];
  starterCode?: string;
  testCases?: TestCase[];
}

export type ProblemUpdateDifficulty = typeof ProblemUpdateDifficulty[keyof typeof ProblemUpdateDifficulty];


export const ProblemUpdateDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface ProblemUpdate {
  title?: string;
  difficulty?: ProblemUpdateDifficulty;
  description?: string;
  constraints?: string;
  hints?: string;
  tags?: string[];
  starterCode?: string;
  testCases?: TestCase[];
}

export interface ProblemListResponse {
  problems: Problem[];
  total: number;
  page: number;
  limit: number;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  running: 'running',
  accepted: 'accepted',
  wrong_answer: 'wrong_answer',
  time_limit_exceeded: 'time_limit_exceeded',
  runtime_error: 'runtime_error',
  compilation_error: 'compilation_error',
} as const;

export interface Submission {
  id: number;
  problemId: number;
  userId: number;
  language: string;
  code: string;
  status: SubmissionStatus;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  memory?: number | null;
  /** @nullable */
  output?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  problemTitle?: string | null;
}

export interface SubmissionInput {
  problemId: number;
  language: string;
  code: string;
  /** @nullable */
  contestId?: number | null;
}

export interface SubmissionListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  limit: number;
}

export interface RunCodeInput {
  language: string;
  code: string;
  input: string;
}

export interface RunResult {
  status: string;
  output: string;
  /** @nullable */
  stderr?: string | null;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  memory?: number | null;
}

export type ContestStatus = typeof ContestStatus[keyof typeof ContestStatus];


export const ContestStatus = {
  upcoming: 'upcoming',
  active: 'active',
  ended: 'ended',
} as const;

export interface Contest {
  id: number;
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  status: ContestStatus;
  problemCount: number;
  participantCount: number;
  problemIds?: number[];
  isRegistered?: boolean;
  createdAt: string;
}

export interface ContestInput {
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  problemIds?: number[];
}

export interface ContestUpdate {
  title?: string;
  description?: string;
  startTime?: string;
  endTime?: string;
  problemIds?: number[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  score: number;
  solvedCount: number;
  /** @nullable */
  runtime?: number | null;
}

export interface DashboardStats {
  totalSolved: number;
  easySolved: number;
  mediumSolved: number;
  hardSolved: number;
  totalProblems: number;
  easyTotal: number;
  mediumTotal: number;
  hardTotal: number;
  submissionsCount: number;
  acceptanceRate: number;
  rank: number;
  streak: number;
}

export interface ProblemStats {
  total: number;
  easy: number;
  medium: number;
  hard: number;
}

export type ListProblemsParams = {
difficulty?: ListProblemsDifficulty;
tag?: string;
status?: ListProblemsStatus;
search?: string;
page?: number;
limit?: number;
};

export type ListProblemsDifficulty = typeof ListProblemsDifficulty[keyof typeof ListProblemsDifficulty];


export const ListProblemsDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ListProblemsStatus = typeof ListProblemsStatus[keyof typeof ListProblemsStatus];


export const ListProblemsStatus = {
  solved: 'solved',
  unsolved: 'unsolved',
  attempted: 'attempted',
} as const;

export type ListSubmissionsParams = {
problemId?: number;
language?: string;
status?: string;
page?: number;
limit?: number;
};

export type ListContestsParams = {
status?: ListContestsStatus;
};

export type ListContestsStatus = typeof ListContestsStatus[keyof typeof ListContestsStatus];


export const ListContestsStatus = {
  upcoming: 'upcoming',
  active: 'active',
  ended: 'ended',
} as const;

export type GetLeaderboardParams = {
page?: number;
limit?: number;
};