# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeLab – LeetCode-style competitive coding platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── PROBLEMS ──────────────────────────────────────────────────────────────
...
  /contests:
    get:
      operationId: listContests
      tags: [contests]
      summary: List contests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, ended]
      responses:
        "200":
          description: Contest list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contest"

    post:
      operationId: createContest
      tags: [contests]
      summary: Create contest (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
  /contests/{id}:
    get:
      operationId: getContest
      tags: [contests]
      summary: Get contest detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateContest
      tags: [contests]
      summary: Update contest (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"

    delete:
      operationId: deleteContest
      tags: [contests]
      summary: Delete contest (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contests/{id}/register:
    post:
      operationId: registerContest
      tags: [contests]
      summary: Register for a contest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /contests/{id}/leaderboard:
    get:
      operationId: getContestLeaderboard
      tags: [contests]
      summary: Get contest leaderboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ── GLOBAL LEADERBOARD ────────────────────────────────────────────────────
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [auth]
      summary: Get public user profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

# ── COMPONENTS ──────────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:          { type: integer }
        username:    { type: string }
        email:       { type: string }
        role:        { type: string, enum: [user, admin] }
        avatarUrl:   { type: ["string", "null"] }
        bio:         { type: ["string", "null"] }
        createdAt:   { type: string }
    UserProfile:
      type: object
      required: [id, username, role, createdAt, solvedCount, rank]
      properties:
        id:          { type: integer }
        username:    { type: string }
        role:        { type: string }
        avatarUrl:   { type: ["string", "null"] }
        bio:         { type: ["string", "null"] }
        createdAt:   { type: string }
        solvedCount: { type: integer }
        rank:        { type: integer }
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:  { type: string, minLength: 3 }
        email:     { type: string }
        password:  { type: string, minLength: 6 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:  { $ref: "#/components/schemas/User" }
        token: { type: string }

    # PROBLEMS
...
    Contest:
      type: object
      required: [id, title, description, startTime, endTime, status, problemCount, participantCount, createdAt]
      properties:
        id:               { type: integer }
        title:            { type: string }
        description:      { type: string }
        startTime:        { type: string }
        endTime:          { type: string }
        status:           { type: string, enum: [upcoming, active, ended] }
        problemCount:     { type: integer }
        participantCount: { type: integer }
        problemIds:       { type: array, items: { type: integer } }
        isRegistered:     { type: boolean }
        createdAt:        { type: string }
    ContestInput:
      type: object
      required: [title, description, startTime, endTime]
      properties:
        title:       { type: string }
        description: { type: string }
        startTime:   { type: string }
        endTime:     { type: string }
        problemIds:  { type: array, items: { type: integer } }
    ContestUpdate:
      type: object
      properties:
        title:       { type: string }
        description: { type: string }
        startTime:   { type: string }
        endTime:     { type: string }
        problemIds:  { type: array, items: { type: integer } }
    LeaderboardEntry:
      type: object
      required: [rank, userId, username, score, solvedCount]
      properties:
        rank:        { type: integer }
        userId:      { type: integer }
        username:    { type: string }
        avatarUrl:   { type: ["string", "null"] }
        score:       { type: integer }
        solvedCount: { type: integer }
        runtime:     { type: ["integer", "null"] }

    # STATS
...